Move find methods to separate files
