+++ src/main/kotlin/cacophony/grammars/FindFollow.kt
@@ -1,0 +1,11 @@
+package cacophony.grammars
+
+import cacophony.automata.DFA
+
+fun <StateType, SymbolType, ResultType> findFollow(
+    automata: Map<SymbolType, DFA<StateType, SymbolType, ResultType>>,
+    nullable: Collection<DFAStateReference<StateType, SymbolType, ResultType>>,
+    first: StateToSymbolsMap<StateType, SymbolType, ResultType>,
+): StateToSymbolsMap<StateType, SymbolType, ResultType> {
+    TODO("Not implemented")
+}
